backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Form Data
python-multipart

# Database & ORM
sqlmodel
psycopg2-binary
alembic
redis

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# Notebook & Data Handling
jupyter-client
pandas

# MLOps & Machine Learning
mlflow
scikit-learn

# LLM Integration
google-generativeai

# Utilities
python-dotenv

# Notebook & Data Handling
jupyter-client
ipykernel
pandas
pandasql
